restructuring feature extractor
